+++ foxplot/functions.py
@@ -1,0 +1,30 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+#
+# SPDX-License-Identifier: Apache-2.0
+# Copyright 2024 Inria
+
+"""Functions that can be applied to frozen series."""
+
+import numpy as np
+from numpy.typing import NDArray
+
+from .frozen_series import FrozenSeries
+
+
+def std(series: FrozenSeries, window_size: int) -> NDArray[float]:
+    """Return the rolling standard deviation of the series.
+
+    Args:
+        window_size: Size of the rolling window in which to compute
+            standard deviations.
+
+    Returns:
+        Array of windowed standard deviations along the series.
+    """
+    label = f"std({series._label}, {window_size})"
+    values = np.std(
+        np.lib.stride_tricks.sliding_window_view(series._values, window_size),
+        axis=-1,
+    )
+    return FrozenSeries(label, values)